The program can now recognize if the ad started or not.
Fixed some syntax.
Improved Sync function.

The program screenshots the instance istantly after hitting the chest and 13 seconds after that hit to check (trought cv2 module) whatever the 2 screeenshots are the same, if so then the ad didn't show up (At this point i just report the problem to the user because implementing a 'retry to watch' would be hard and longer and will slow down the other instances (UNLESS I CONSIDER EACH INSTANCE AS A TRHEAD, so each instance is independent, it hasn't synced actions).

-> Need to code the process of closing all the different ads :
 - 5 sec ad
 - 30 sec ad
 - interactive ad : to skip the ad you have to click on a '>>' symbol and then on 2 X buttons which come one next to the one before every 5 seconds.