Added an Ad-Closing automation script
Improved CompareScreenshots() function by comparing the similarity of 2 pictures instead of the size in bytes.

Improved some typos/outputs.

Worked on GetImageMatches() to check whatever a smaller image (X button img) is found inside the main image (linked the yt video in the py comments that explains it).

Finally coded the CloseAd() function which get a screenshot of the instance after 30 seconds (time to let any ad end) and look for any match with any of the many X-Buttons.png images formats.
There's a problem with interactive ads where the 'skip' button is a '>>' button but with the background transparent, so that is not possible to look for a match for that button. -> Solved this issue by default-clicking on X,Y coordinates where the 'skip' buttons usually are.
-> The program knows if the ad has been closed by re-screenshotting the screen and look for a match of the 'item-bag' image.

Then the program kills all instance's psg2 processes and start the next cloned app game (done in the for loop in line 367.